requirements.txt
python-telegram-bot[webhooks]>=22.0
openai>=1.0.0
anthropic>=0.40.0
python-dotenv>=1.0.0
yfinance>=0.2.50
pandas>=2.2.0
tenacity>=9.0.0
diskcache>=5.6.3
redis>=5.2.0
PyYAML>=6.0.2
matplotlib>=3.9.0
transformers>=4.46.0
torch>=2.4.0
sentry-sdk>=2.18.0
openpyxl>=3.1.5
pytest>=8.3.0
